Logging refactoring
Use a defined logger (from `logging.getLogger(__name__)` instead of calling `logging` because this is probably a better practice in the long run.

Paramaters dict (p) is no longer explicitly defined as a global within a function, but assigned at the top level of scope before the call to `main`.